+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## 0.2.5-dev
+## Main/Unreleased
+
+### Models
+
+### Tasks
+
+### Metrics
+
+### General
+
+## 0.2.5
 
 ### Models
 - Move `aleph_alpha.py` to use `/completions` endpoint instead of `/evaluate`. `/evaluate` was just available for model deployed in the luminous workers and is not supported in vllm.
@@ -14,6 +24,7 @@
 
 ### General
 - Updated image urls to be absolute so the pypi page can display them correctly
+- Added `llm_judge_prompt` and `llm_judge_response` to MTBENCH metric results
 
 ## 0.2.4
 
@@ -32,7 +43,7 @@
 
 - Added documentation on `SQUAD` and `SQUAD2` benchmark classes
 - Updated documentation on lists of available tasks
-- added `.vscode/launch.json`
+- Added `.vscode/launch.json`
 - Added verbosity levels (0 is critical, 1 is info, 2 is debug) for minimal output
 - Modified the Hendrycks Math task to use the same query template as MATH500 to encourage boxed answer formatting.